Changes the way bookmark function works
Instead of setting a new variable containing the path to the target location,
creates symbolic links in ~/bookmarks. The rationale behind this change is
that it makes it easier to check what shortcuts are defined, with a simple
ls -l ~/bookmarks.
TODO: implement something like a --show option for bookmark, and maybe make
~/bookmarks hidden (optional)